Addd Templates and Inventories Tabs. Refactor the platform component to use React hooks
