+++ src/smart-components/platform/platform.js
@@ -1,127 +1,52 @@
-import React, { Component, Fragment } from 'react';
+import React, { Fragment, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import debouncePromise from 'awesome-debounce-promise';
+import { Route, Switch } from 'react-router-dom';
+import { scrollToTop } from '../../helpers/shared/helpers';
+import { fetchSelectedPlatform } from '../../redux/actions/platform-actions';
+import PlatformTemplates from './platform_templates';
+import PlatformInventories from './platform_inventories';
+import AppTabs from './../../presentational-components/shared/app-tabs';
 
-import ToolbarRenderer from '../../toolbar/toolbar-renderer';
-import ContentGallery from '../content-gallery/content-gallery';
-import { scrollToTop, filterServiceOffering } from '../../helpers/shared/helpers';
-import PlatformItem from '../../presentational-components/platform/platform-item';
-import createPlatformsToolbarSchema from '../../toolbar/schemas/platforms-toolbar.schema';
-import { defaultSettings, getCurrentPage, getNewPage } from '../../helpers/shared/pagination';
-import { fetchSelectedPlatform, fetchPlatformItems } from '../../redux/actions/platform-actions';
+const Platform = (props) => {
 
-class Platform extends Component {
-  state = {
-    filterValue: ''
-  };
+  const tabItems = [{ eventKey: 0, title: 'Templates', name: `/platforms/detail/${props.match.params.id}/platform-templates` },
+    { eventKey: 1, title: 'Inventories', name: `/platforms/detail/${props.match.params.id}/platform-inventories` }];
 
-  fetchData(apiProps, pagination) {
-    this.props.fetchSelectedPlatform(apiProps);
-    this.props.fetchPlatformItems(apiProps, pagination);
-  }
-
-  componentDidMount() {
-    this.fetchData(this.props.match.params.id, defaultSettings);
+  useEffect(() => {
+    this.fetchData(props.match.params.id, defaultSettings);
     scrollToTop();
-  }
+  }, [ props.match.params.id ]);
 
-  componentDidUpdate(prevProps) {
-    if (prevProps.match.params.id !== this.props.match.params.id) {
-      this.fetchData(this.props.match.params.id, defaultSettings);
-      scrollToTop();
-    }
-  }
+  return (
+    <Fragment>
+      <AppTabs tabItems={ tabItems }/>
+      <Switch>
+        <Route path={ `/platforms/detail/:id/platform-templates` } component={ PlatformTemplates }/>
+        <Route path={ `/platforms/detail/:id/platform-inventories` } component={ PlatformInventories }/>
+      </Switch>
+    </Fragment>
+  );
+};
 
-  handleOnPerPageSelect = limit => this.props.fetchPlatformItems(this.props.match.params.id, {
-    offset: this.props.paginationCurrent.offset,
-    limit
-  });
-
-  handleSetPage = (number, debounce) => {
-    const options = {
-      offset: getNewPage(number, this.props.paginationCurrent.limit),
-      limit: this.props.paginationCurrent.limit
-    };
-    const request = () => this.props.fetchPlatformItems(this.props.match.params.id, options);
-    if (debounce) {
-      return debouncePromise(request, 250)();
-    }
-
-    return request();
-  }
-
-  handleFilterChange = filterValue => this.setState({ filterValue });
-
-  render() {
-    let filteredItems = {
-      items: this.props.platformItems
-      .filter(item => filterServiceOffering(item, this.state.filterValue))
-      .map(data => <PlatformItem key={ data.id } { ...data } />),
-      isLoading: this.props.isPlatformDataLoading
-    };
-
-    let title = this.props.platform ? this.props.platform.name : '';
-    return (
-      <Fragment>
-        <ToolbarRenderer schema={ createPlatformsToolbarSchema({
-          onFilterChange: this.handleFilterChange,
-          searchValue: this.state.filterValue,
-          title,
-          pagination: {
-            itemsPerPage: this.props.paginationCurrent.limit,
-            numberOfItems: this.props.paginationCurrent.count || 50,
-            onPerPageSelect: this.handleOnPerPageSelect,
-            page: getCurrentPage(this.props.paginationCurrent.limit, this.props.paginationCurrent.offset),
-            onSetPage: this.handleSetPage,
-            direction: 'down'
-          }
-        }) }/>
-        <ContentGallery { ...filteredItems }/>
-      </Fragment>
-    );
-  }
-}
-
-const mapStateToProps = ({ platformReducer: { selectedPlatform, platformItems, isPlatformDataLoading }}) => {
-  const platformItemsData = selectedPlatform && platformItems[selectedPlatform.id];
+const mapStateToProps = ({ platformReducer: { selectedPlatform }}) => {
   return {
-    paginationLinks: platformItemsData && platformItemsData.links,
-    paginationCurrent: platformItemsData && platformItemsData.meta,
     platform: selectedPlatform,
-    platformItems: platformItemsData && platformItemsData.data,
-    isPlatformDataLoading: !selectedPlatform || isPlatformDataLoading
+    isPlatformDataLoading: selectedPlatform
   };
 };
 
 const mapDispatchToProps = dispatch => bindActionCreators({
-  fetchSelectedPlatform,
-  fetchPlatformItems
+  fetchSelectedPlatform
 }, dispatch);
 
 Platform.propTypes = {
-  filteredItems: PropTypes.object,
-  isPlatformDataLoading: PropTypes.bool,
   match: PropTypes.object,
-  fetchPlatformItems: PropTypes.func.isRequired,
   fetchSelectedPlatform: PropTypes.func,
   platform: PropTypes.shape({
     name: PropTypes.string
-  }),
-  platformItems: PropTypes.array,
-  paginationCurrent: PropTypes.shape({
-    limit: PropTypes.number.isRequired,
-    offset: PropTypes.number.isRequired,
-    count: PropTypes.number.isRequired
   })
-};
-
-Platform.defaultProps = {
-  platformItems: [],
-  paginationCurrent: {
-    limit: 50
-  }
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Platform);
